requirements.txt
pyautogui
pillow
customtkinter
mouse
pyinstaller